fix(apps/swap): button not disabled when there was a min output violation error
