+++ apps/evm/ui/swap/simple/simple-swap-trade-button.tsx
@@ -19,15 +19,28 @@ import { APPROVE_TAG_SWAP } from 'lib/constants'
 import { warningSeverity } from 'lib/swap/warningSeverity'
 import React, { FC, useEffect, useState } from 'react'
 
+import { usePersistedSlippageError } from '../../../lib/hooks'
 import { useDerivedStateSimpleSwap, useSimpleSwapTrade } from './derivedstate-simple-swap-provider'
 import { SimpleSwapTradeReviewDialog } from './simple-swap-trade-review-dialog'
 
 export const SimpleSwapTradeButton: FC = () => {
+  return (
+    <>
+      <SimpleSwapTradeReviewDialog>
+        {({ error, isSuccess }) => <_SimpleSwapTradeButton error={error} isSuccess={isSuccess} />}
+      </SimpleSwapTradeReviewDialog>
+    </>
+  )
+}
+
+export const _SimpleSwapTradeButton: FC<{ error: Error | null; isSuccess: boolean }> = ({ error, isSuccess }) => {
+  const { isSlippageError } = usePersistedSlippageError({ isSuccess, error })
   const { data: trade } = useSimpleSwapTrade()
+  const [checked, setChecked] = useState(false)
+
   const {
     state: { swapAmount, swapAmountString, chainId, token0, token1 },
   } = useDerivedStateSimpleSwap()
-  const [checked, setChecked] = useState(false)
 
   const isWrap = token0?.isNative && token1?.wrapped.address === Native.onChain(chainId).wrapped.address
   const isUnwrap = token1?.isNative && token0?.wrapped.address === Native.onChain(chainId).wrapped.address
@@ -41,61 +54,58 @@ export const SimpleSwapTradeButton: FC = () => {
 
   return (
     <>
-      <SimpleSwapTradeReviewDialog>
-        {(error) => (
-          <div>
-            <Checker.Connect>
-              <Checker.Network chainId={chainId}>
-                <Checker.Amounts chainId={chainId} amounts={[swapAmount]}>
-                  <Checker.ApproveERC20
-                    id="approve-erc20"
-                    amount={swapAmount}
-                    contract={
-                      isRouteProcessor3_2ChainId(chainId)
-                        ? ROUTE_PROCESSOR_3_2_ADDRESS[chainId]
-                        : isRouteProcessor3_1ChainId(chainId)
-                        ? ROUTE_PROCESSOR_3_1_ADDRESS[chainId]
-                        : isRouteProcessor3ChainId(chainId)
-                        ? ROUTE_PROCESSOR_3_ADDRESS[chainId]
-                        : isRouteProcessorChainId(chainId)
-                        ? ROUTE_PROCESSOR_ADDRESS[chainId]
-                        : undefined
-                    }
-                  >
-                    <Checker.Success tag={APPROVE_TAG_SWAP}>
-                      <DialogTrigger asChild>
-                        <Button
-                          size="xl"
-                          disabled={Boolean(
-                            error ||
-                              !trade?.amountOut?.greaterThan(ZERO) ||
-                              trade?.route?.status === 'NoWay' ||
-                              +swapAmountString === 0 ||
-                              (!checked && warningSeverity(trade?.priceImpact) > 3)
-                          )}
-                          color={warningSeverity(trade?.priceImpact) >= 3 ? 'red' : 'blue'}
-                          fullWidth
-                          testId="swap"
-                        >
-                          {!checked && warningSeverity(trade?.priceImpact) >= 3
-                            ? 'Price impact too high'
-                            : trade?.route?.status === 'NoWay'
-                            ? 'No trade found'
-                            : isWrap
-                            ? 'Wrap'
-                            : isUnwrap
-                            ? 'Unwrap'
-                            : 'Swap'}
-                        </Button>
-                      </DialogTrigger>
-                    </Checker.Success>
-                  </Checker.ApproveERC20>
-                </Checker.Amounts>
-              </Checker.Network>
-            </Checker.Connect>
-          </div>
-        )}
-      </SimpleSwapTradeReviewDialog>
+      <div>
+        <Checker.Connect>
+          <Checker.Network chainId={chainId}>
+            <Checker.Amounts chainId={chainId} amounts={[swapAmount]}>
+              <Checker.ApproveERC20
+                id="approve-erc20"
+                amount={swapAmount}
+                contract={
+                  isRouteProcessor3_2ChainId(chainId)
+                    ? ROUTE_PROCESSOR_3_2_ADDRESS[chainId]
+                    : isRouteProcessor3_1ChainId(chainId)
+                    ? ROUTE_PROCESSOR_3_1_ADDRESS[chainId]
+                    : isRouteProcessor3ChainId(chainId)
+                    ? ROUTE_PROCESSOR_3_ADDRESS[chainId]
+                    : isRouteProcessorChainId(chainId)
+                    ? ROUTE_PROCESSOR_ADDRESS[chainId]
+                    : undefined
+                }
+              >
+                <Checker.Success tag={APPROVE_TAG_SWAP}>
+                  <DialogTrigger asChild>
+                    <Button
+                      size="xl"
+                      disabled={Boolean(
+                        isSlippageError ||
+                          error ||
+                          !trade?.amountOut?.greaterThan(ZERO) ||
+                          trade?.route?.status === 'NoWay' ||
+                          +swapAmountString === 0 ||
+                          (!checked && warningSeverity(trade?.priceImpact) > 3)
+                      )}
+                      color={warningSeverity(trade?.priceImpact) >= 3 ? 'red' : 'blue'}
+                      fullWidth
+                      testId="swap"
+                    >
+                      {!checked && warningSeverity(trade?.priceImpact) >= 3
+                        ? 'Price impact too high'
+                        : trade?.route?.status === 'NoWay'
+                        ? 'No trade found'
+                        : isWrap
+                        ? 'Wrap'
+                        : isUnwrap
+                        ? 'Unwrap'
+                        : 'Swap'}
+                    </Button>
+                  </DialogTrigger>
+                </Checker.Success>
+              </Checker.ApproveERC20>
+            </Checker.Amounts>
+          </Checker.Network>
+        </Checker.Connect>
+      </div>
       {warningSeverity(trade?.priceImpact) > 3 && (
         <div className="flex items-start px-4 py-3 mt-4 rounded-xl bg-red/20">
           <input
